+++ src/services/authService.js
@@ -1,0 +1,39 @@
+const bcrypt = require('bcryptjs');
+const { PrismaClient } = require('@prisma/client');
+const prisma = new PrismaClient();
+
+const registerUser = async ({ username, password, email, role }) => {
+  // Check if user already exists
+  const existingUser = await prisma.user.findUnique({
+    where: { email },
+  });
+
+  if (existingUser) {
+    throw new Error('Email already exists');
+  }
+
+  // Hash password
+  const salt = await bcrypt.genSalt(10);
+  const hashedPassword = await bcrypt.hash(password, salt);
+
+  // Create user
+  const newUser = await prisma.user.create({
+    data: {
+      username,
+      email,
+      password: hashedPassword,
+      role,
+    },
+  });
+
+  return {
+    id: newUser.id,
+    username: newUser.username,
+    email: newUser.email,
+    role: newUser.role,
+  };
+};
+
+module.exports = {
+  registerUser,
+};
